Do not pass CardHeader property titleColor to the DOM
Using the titleColor property on CardHeader, react gives a warning
"Warning: Unknown prop `titleColor` on <div> tag. Remove this prop from the element. For details, see https://fb.me/react-unknown-prop"

Note that the titleColor works properly, the header text is displayed in the given color, the only problem is the warning.

Closes #5047.