[core] Fix component wrong propType
Following #6238, we are expecting an input for the `React.createElement` API.
Hence, using `PropTypes.element` do not make any sense.
It would if we were using `React.cloneElement` but we don't.
I should have seen is earlier.
I have added a test case and a test security to prevent this type of regression to happen again.