[github] Add Support Requests bot config file (#9751)
This is a first pass at the wording for an automated response when an issue is tagged "support".

This feature will only be used for questions that are clearly how-to questions, or other non-issues.  Other questions that a grey-are should be tagged "question" until they are confirmed as an issue or tagged "support". (Looking at you @mbrookes!)

https://probot.github.io/apps/support/