[core] Ponyfill global (#13426)
* Polyfill global

`global` is commonjs specific object. Distributing it makes user build
harder. Well, webpack handles this case but this package is web
specific, not node so it's better to provide own polyfill.

This should simplify configuration of bundlers like rollup which handles
only esm out of the box.

JSS already do this for future v10 and would be great to fix this at
least for one package.

* Format files

* Disable eslint for global polyfill

* copy core-js ponyfill